Rename: styles/components 폴더 밑으로 컴포넌트 css파일 이동
