+++ src/components/Footer.tsx
@@ -1,4 +1,4 @@
-import styles from "@/styles/footer.module.css";
+import styles from "@/styles/components/footer.module.css";
 
 export default function Footer() {
   return (
