+++ src/styles/components/commonframe.module.css
